soc-bot/requirements.txt
apscheduler>=3.10
requests>=2.31
redis>=5.0
python-dotenv>=1.0
pyyaml>=6.0
urllib3>=2.0
